[refactoring/daengle-58] 코드 컨벤션 통일
