+++ daengle-domain/src/main/java/ddog/domain/payment/port/OrderPersist.java
@@ -6,6 +6,6 @@ import java.util.Optional;
 
 public interface OrderPersist {
     Order save(Order order);
-    Optional<Order> findBy(String orderUid);
+    Optional<Order> findByOrderUid(String orderUid);
     void delete(Order order);
 }
