+++ daengle-domain/src/main/java/ddog/domain/review/Review.java
@@ -23,4 +23,22 @@ public abstract class Review {
     private LocalDateTime createTime;
     private LocalDateTime modifiedTime;
     private List<String> imageUrlList;
+
+    public static void validateStarRating(Long starRating) {
+        if (starRating == null || starRating < 1 || starRating > 5) {
+            throw new IllegalArgumentException("Star rating must be between 1 and 5.");
+        }
+    }
+
+    public static void validateContent(String content) {
+        if (content != null && content.length() > 400) {
+            throw new IllegalArgumentException("Content must be 400 characters or less.");
+        }
+    }
+
+    public static void validateImageUrlList(List<String> imageUrlList) {
+        if (imageUrlList != null && imageUrlList.size() > 10) {
+            throw new IllegalArgumentException("Image URL list must contain 10 items or less.");
+        }
+    }
 }
