+++ daengle-domain/src/main/java/ddog/domain/review/port/CareReviewPersist.java
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Pageable;
 import java.util.Optional;
 
 public interface CareReviewPersist {
-    Optional<CareReview> findBy(Long careReviewId);
+    Optional<CareReview> findByReviewId(Long careReviewId);
     CareReview save(CareReview careReview);
     void delete(CareReview careReview);
     Page<CareReview> findByReviewerId(Long userId, Pageable pageable);
