+++ daengle-domain/src/main/java/ddog/domain/user/port/UserPersist.java
@@ -9,7 +9,5 @@ public interface UserPersist {
 
     Optional<User> findByAccountId(Long accountId);
 
-    Optional<User> findBy(Long id);
-
     Boolean hasNickname(String nickname);
 }
